# Optimised vectorised global alignment.