--- a/include/seqan3/alignment/matrix/detail/matrix_coordinate.hpp
+++ b/include/seqan3/alignment/matrix/detail/matrix_coordinate.hpp
@@ -16,15 +16,20 @@
 
 #include <seqan3/core/concept/core_language.hpp>
 #include <seqan3/core/detail/strong_type.hpp>
+#include <seqan3/core/simd/concept.hpp>
+#include <seqan3/core/simd/simd_algorithm.hpp>
 #include <seqan3/std/concepts>
 
 namespace seqan3::detail
 {
 /*!\brief A strong type for designated initialisation of the column index of a marix.
  * \ingroup alignment_matrix
- * \tparam index_type The type of the index to store; must model seqan3::arithmetic.
+ * \tparam index_type The type of the index to store; must model seqan3::arithmetic or seqan3::simd::simd_concept.
  */
-template <std::integral index_type>
+template <typename index_type>
+//!\cond
+    requires std::integral<index_type> || simd_concept<index_type>
+//!\endcond
 struct column_index_type : detail::strong_type<index_type, column_index_type<index_type>>
 {
     //!!\brief Import base class constructor.
@@ -42,13 +47,20 @@ column_index_type(index_type) -> column_index_type<std::ptrdiff_t>;
 //!\brief Deduces an unsigned integral type to size_t.
 template <std::unsigned_integral index_type>
 column_index_type(index_type) -> column_index_type<size_t>;
+
+//!\brief Deduces the template argument from a simd vector index type.
+template <simd_concept index_type>
+column_index_type(index_type) -> column_index_type<index_type>;
 //!\}
 
 /*!\brief A strong type for designated initialisation of the row index of a marix.
  * \ingroup alignment_matrix
- * \tparam index_type The type of the index to store; must model seqan3::arithmetic.
+ * \tparam index_type The type of the index to store; must model seqan3::arithmetic or seqan3::simd::simd_concept
  */
-template <std::integral index_type>
+template <typename index_type>
+//!\cond
+    requires std::integral<index_type> || simd_concept<index_type>
+//!\endcond
 struct row_index_type : detail::strong_type<index_type, row_index_type<index_type>>
 {
     //!!\brief Import base class constructor.
@@ -66,11 +78,20 @@ row_index_type(index_type) -> row_index_type<std::ptrdiff_t>;
 //!\brief Deduces an unsigned integral type to size_t.
 template <std::unsigned_integral index_type>
 row_index_type(index_type) -> row_index_type<size_t>;
+
+//!\brief Deduces the template argument from a simd vector index type.
+template <simd_concept index_type>
+row_index_type(index_type) -> row_index_type<index_type>;
 //!\}
 
-//!\brief A representation of a location or offset within a two-dimensional matrix.
-//!\ingroup alignment_matrix
-template <std::integral index_t>
+/*!\brief A representation of a location or offset within a two-dimensional matrix.
+ * \ingroup alignment_matrix
+ * \tparam index_t The underlying index type; must model seqan3::arithmetic or seqan3::simd::simd_concept.
+ */
+template <typename index_t>
+//!\cond
+    requires std::integral<index_t> || simd_concept<index_t>
+//!\endcond
 struct matrix_index
 {
     /*!\name Constructors, destructor and assignment
@@ -87,9 +108,33 @@ struct matrix_index
      * \param row_idx The row index to set.
      * \param col_idx The column index to set.
      */
-    constexpr matrix_index(row_index_type<index_t> const row_idx,
-                           column_index_type<index_t> const col_idx) noexcept
-        : row{row_idx.get()}, col{col_idx.get()}
+    constexpr matrix_index(row_index_type<index_t> const row_idx, column_index_type<index_t> const col_idx) noexcept :
+        row{row_idx.get()},
+        col{col_idx.get()}
+    {}
+
+    /*!\brief Construction from strongly typed row index and column index over a scalar type when the index is a simd
+     *        vector.
+     *
+     * \tparam scalar_index_t The type of the scalar index type; must model seqan3::arithmetic.
+     *
+     * \param row_idx The row index to set.
+     * \param col_idx The column index to set.
+     *
+     * \details
+     *
+     * This constructor initialises the row and col index which is represented as simd vectors. The vectors are
+     * initialised with the scalar types of the given strong types over the scalar value. This constructor is only
+     * available if `index_t` is a simd vector type.
+     */
+    template <seqan3::arithmetic scalar_index_t>
+    constexpr matrix_index(row_index_type<scalar_index_t> const row_idx,
+                           column_index_type<scalar_index_t> const col_idx) noexcept
+    //!\cond
+        requires simd_concept<index_t>
+    //!\endcond
+        : row{simd::fill<index_t>(static_cast<typename simd_traits<index_t>::scalar_type>(row_idx.get()))},
+          col{simd::fill<index_t>(static_cast<typename simd_traits<index_t>::scalar_type>(col_idx.get()))}
     {}
 
     /*!\brief Construction from other matrix_index with different integral type.
@@ -130,11 +175,21 @@ template <std::integral row_index_t, std::integral col_index_t>
 matrix_index(row_index_type<row_index_t>, column_index_type<col_index_t>) ->
     matrix_index<std::common_type_t<row_index_t, col_index_t>>;
 
+//!\brief Deduces the index type from the simd vector index type.
+template <simd_concept index_t>
+matrix_index(row_index_type<index_t>, column_index_type<index_t>) -> matrix_index<index_t>;
 //!\}
 
 //!\brief A coordinate type to access an element inside of a two-dimensional matrix.
 //!\ingroup alignment_matrix
 using matrix_coordinate = matrix_index<size_t>;
+
+/*!\brief A coordinate type to access an element inside of a two-dimensional simd vector matrix.
+ * \ingroup alignment_matrix
+ * \tparam index_t The underlying index type; must model seqan3::simd::simd_concept.
+ */
+template <simd_concept index_t>
+using simd_matrix_coordinate = matrix_index<index_t>;
 
 //!\brief An offset type to move a matrix iterator in two-dimensional space.
 //!\ingroup alignment_matrix