… include/seqan3/alignment/pairwise/alignment_configurator.hpp
… include/seqan3/alignment/pairwise/detail/pairwise_alignment_algorithm.hpp
… include/seqan3/alignment/pairwise/detail/policy_affine_gap_recursion.hpp
… include/seqan3/alignment/pairwise/detail/policy_affine_gap_recursion_simd.hpp
… include/seqan3/alignment/pairwise/detail/policy_optimum_tracker.hpp
… include/seqan3/alignment/pairwise/detail/policy_optimum_tracker_simd.hpp
… include/seqan3/alignment/pairwise/detail/type_traits.hpp
… include/seqan3/core/simd/all.hpp
… include/seqan3/core/simd/detail/simd_algorithm_avx2.hpp
… include/seqan3/core/simd/simd_algorithm.hpp
… include/seqan3/core/simd/view_iota_simd.hpp
… include/seqan3/core/simd/view_to_simd.hpp
… test/performance/alignment/global_affine_alignment_simd_benchmark.cpp
… test/performance/core/simd/CMakeLists.txt
… test/performance/core/simd/simd_iota_benchmark.cpp
… test/snippet/core/simd/view_iota_simd.cpp
… test/unit/alignment/matrix/detail/CMakeLists.txt
… test/unit/alignment/matrix/detail/coordinate_matrix_simd_test.cpp
… test/unit/alignment/matrix/detail/score_matrix_single_column_simd_test.cpp
… test/unit/alignment/pairwise/pairwise_alignment_collection_test_template.hpp
… test/unit/core/simd/CMakeLists.txt
… test/unit/core/simd/view_iota_simd_test.cpp