Enables the vectorised global alignment.

Micro benchmark with AVX2:

### Seqan3 new
```console
Run on (4 X 3600 MHz CPU s)
CPU Caches:
  L1 Data 32K (x4)
  L1 Instruction 32K (x4)
  L2 Unified 262K (x4)
  L3 Unified 6291K (x1)
Load Average: 3.50, 3.48, 3.09
--------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                            Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------------------------------------------------
seqan3_affine_dna4_accelerated/simd_with_score/0/real_time                     2795783 ns      2786265 ns          253 CUPS=8.35123G/s cells=23.3482M total=-18.004M
seqan3_affine_dna4_accelerated/simd_with_score/8/real_time                     2987575 ns      2980251 ns          235 CUPS=7.79106G/s cells=23.2764M total=-16.8248M
seqan3_affine_dna4_accelerated/simd_with_score/16/real_time                    3239686 ns      3229834 ns          211 CUPS=7.16485G/s cells=23.2119M total=-15.1144M
seqan3_affine_dna4_accelerated/simd_with_score/24/real_time                    3511635 ns      3498046 ns          197 CUPS=6.59069G/s cells=23.1441M total=-14.2201M
seqan3_affine_dna4_accelerated/simd_with_score/32/real_time                    3779830 ns      3771973 ns          183 CUPS=6.10703G/s cells=23.0835M total=-13.3112M
seqan3_affine_dna4_accelerated/simd_with_score/40/real_time                    4113648 ns      4079439 ns          171 CUPS=5.59517G/s cells=23.0166M total=-12.6039M
seqan3_affine_dna4_accelerated/simd_with_score/48/real_time                    4402465 ns      4390226 ns          159 CUPS=5.2142G/s cells=22.9553M total=-11.8838M
seqan3_affine_dna4_accelerated/simd_with_score/56/real_time                    4735617 ns      4722425 ns          134 CUPS=4.83419G/s cells=22.8929M total=-10.1945M
seqan3_affine_dna4_accelerated/simd_with_score/64/real_time                    5010244 ns      5000563 ns          119 CUPS=4.55656G/s cells=22.8295M total=-9.26189M
```

### Seqan3 old
```console
--------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                            Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------------------------------------------------
seqan3_affine_dna4_accelerated/simd_with_score/0/real_time              6307865 ns      6294114 ns          105 CUPS=3.70145G/s cells=23.3482M total=-7.47201M
seqan3_affine_dna4_accelerated/simd_with_score/8/real_time              6898890 ns      6875780 ns           91 CUPS=3.37393G/s cells=23.2764M total=-6.51515M
seqan3_affine_dna4_accelerated/simd_with_score/16/real_time             7481224 ns      7467034 ns           89 CUPS=3.10268G/s cells=23.2119M total=-6.37525M
seqan3_affine_dna4_accelerated/simd_with_score/24/real_time             8088074 ns      8068741 ns           81 CUPS=2.86151G/s cells=23.1441M total=-5.84682M
seqan3_affine_dna4_accelerated/simd_with_score/32/real_time             8720717 ns      8704230 ns           74 CUPS=2.64698G/s cells=23.0835M total=-5.38269M
seqan3_affine_dna4_accelerated/simd_with_score/40/real_time             9466366 ns      9434394 ns           71 CUPS=2.4314G/s cells=23.0166M total=-5.2332M
seqan3_affine_dna4_accelerated/simd_with_score/48/real_time            10157369 ns     10136308 ns           65 CUPS=2.25997G/s cells=22.9553M total=-4.85816M
seqan3_affine_dna4_accelerated/simd_with_score/56/real_time            10849907 ns     10826919 ns           62 CUPS=2.10996G/s cells=22.8929M total=-4.71684M
seqan3_affine_dna4_accelerated/simd_with_score/64/real_time            11632689 ns     11603672 ns           58 CUPS=1.96253G/s cells=22.8295M total=-4.5142M
```

### SeqAn2

```console
--------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                            Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------------------------------------------------
seqan2_affine_dna4_accelerated/simd_with_score/0/real_time                     2913552 ns      2839759 ns          241 CUPS=8.01366G/s cells=23.3482M total=-17.15M
seqan2_affine_dna4_accelerated/simd_with_score/8/real_time                     4867336 ns      4811969 ns          129 CUPS=4.78216G/s cells=23.2764M total=-9.23575M
seqan2_affine_dna4_accelerated/simd_with_score/16/real_time                    5170272 ns      5116564 ns          117 CUPS=4.48949G/s cells=23.2119M total=-8.38094M
seqan2_affine_dna4_accelerated/simd_with_score/24/real_time                    5620308 ns      5575000 ns          111 CUPS=4.11794G/s cells=23.1441M total=-8.01231M
seqan2_affine_dna4_accelerated/simd_with_score/32/real_time                    6079766 ns      6050437 ns          103 CUPS=3.79678G/s cells=23.0835M total=-7.49212M
seqan2_affine_dna4_accelerated/simd_with_score/40/real_time                    6523690 ns      6501835 ns           97 CUPS=3.52815G/s cells=23.0166M total=-7.14958M
seqan2_affine_dna4_accelerated/simd_with_score/48/real_time                    7088224 ns      7049178 ns           90 CUPS=3.23852G/s cells=22.9553M total=-6.72669M
seqan2_affine_dna4_accelerated/simd_with_score/56/real_time                    7534605 ns      7501449 ns           78 CUPS=3.03837G/s cells=22.8929M total=-5.93408M
seqan2_affine_dna4_accelerated/simd_with_score/64/real_time                    8076749 ns      8005354 ns           79 CUPS=2.82657G/s cells=22.8295M total=-6.14865M

```