@rrahn 2020-07-03
> Can we pull out all the simd related changes into a new PR? I mean changes to /core/simd.

https://github.com/seqan/seqan3/pull/1823 is the last extracted PR for this. Sorry, if I forgot to remove you from the reviews.